separate vird list collection data source and bird data source, make collection view private
